exec

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: lumarque <lumarque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/16 16:30:31 by lumarque          #+#    #+#             */
-/*   Updated: 2023/04/16 16:32:29 by lumarque         ###   ########.fr       */
+/*   Updated: 2024/08/01 00:09:40 by lumarque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
